Refactor code into better parts
